lightrag_api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
lightrag-hku>=1.4.9.10
nest_asyncio>=1.6.0
numpy>=1.24.0
